Move Cfd types into the projection actor
Move the logic of creating Cfd representation suitable for UI into the
projection actor

For the time being, keep the logic as free functions that are still used in the
"old" way (inside ToSseEvent)
